pyproject.toml
[tool.poetry]
name = "uema-health"
version = "0.1.0"
description = ""
authors = ["Uemak <kanzi1234567890@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit-aggrid = "^0.3.3"
requests = "^2.28.1"
pandas = "^1.5.0"
plotly = "^5.10.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
